refactor: change dependencies
