+++ Core/IUnitOfWork.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
-using GoodNews.DAL.Repository;
+using GoodNews.Core;
 using GoodNews.Data.Entities;
 
 namespace Core
