+++ GoodNews.BL/GoodNews.BL.csproj
@@ -15,6 +15,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Core\Core.csproj" />
+    <ProjectReference Include="..\GoodNews.DAL\GoodNews.DAL.csproj" />
     <ProjectReference Include="..\Services\Services.csproj" />
   </ItemGroup>
