+++ GoodNews.BL/Startup.cs
@@ -1,4 +1,5 @@
 ﻿using Core;
+using GoodNews.Core;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
