+++ GoodNews.DAL/GoodNews.DAL.csproj
@@ -16,6 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Core\Core.csproj" />
     <ProjectReference Include="..\GoodNews.Data\GoodNews.Data.csproj" />
   </ItemGroup>
