+++ GoodNews.DAL/Repository/EntityFrameworkRepository.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using GoodNews.Core;
 using GoodNews.Data.Entities;
 using Microsoft.EntityFrameworkCore;
