+++ Services/Parsers/IOnlinerParser.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Core;
+﻿using Core;
 
 namespace Services.Parsers
 {
