+++ Services/Services.csproj
@@ -13,6 +13,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Core\Core.csproj" />
+    <ProjectReference Include="..\GoodNews.DAL\GoodNews.DAL.csproj" />
   </ItemGroup>
 
 </Project>
