+++ Services/UnitOfWork.cs
@@ -2,8 +2,8 @@
 using System.Linq;
 using System.Threading.Tasks;
 using Core;
+using GoodNews.Core;
 using GoodNews.DAL;
-using GoodNews.DAL.Repository;
 using GoodNews.Data.Entities;
 
 namespace Services
